requirements.txt
appdirs==1.4.3
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
config-checker==0.0.1
coverage==5.1
db-ops==0.0.3
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
numpy==1.18.4
openpyxl==3.0.3
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
